requirements.txt
amqp==5.2.0
asgiref ==3.8.1
asttokens==2.4.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==5.1.1
Django==5.0.6
django-ckeditor==6.7.1
django-ckeditor-youtube-plugin==0.0.1
django-extensions==3.2.3
django-filter==24.2
django-js-asset==2.2.0
executing==2.0.1
ipython==8.25.0
jedi==0.19.1
kombu==5.3.7
matplotlib-inline==0.1.7
parso==0.8.4
pillow==10.3.0
pip==23.2.1
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.7
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
tzdata==2024.1
Unidecode==1.3.8
vine==5.1.0
wcwidth ==0.2.13
